Now I'm reminded why I hate the Bourne language.
The test command has serious problems because it has no access to the
shell's syntax tree. This means it has no idea if a '!' in the first
argument is an actual operator or a value from a variable.

This can be worked around by counting arguments and looking at the second
and third arguments until you get to the "-a" and "-o" operators. If you
use these operators you can get sets of arguments that are undecidably
ambiguous.

Another way is to add lots of extra quoting (for example prefixing
every value by an 'X') to ensure that the value arguments can never be
mis-interpreted as operators. This is very messy.

The ksh command '[[' helps here because it's defined to be part of the
shell and has access to flags that say if an argument was quoted. If it
was it's obviously not an operator.

PS: I think I might start adding commit messages :-)