Change defaults to ec256+sha256 and other fixes
For versions that don't do that try for rsa-2048+sha256.
For really old versions error so the use has to choose a digest.
Also limit the end date to 2038 for old versions of openssl.

Using here strings increases the required bash version
so replace it by a simple pipe.

Fix for -suca on very old versions

Comments etc.