Change whatssl for new version of openssl verify.
The old "-CApath /dev/null" hack no longer works to forbid use of the
default CA store, but instead we have a nice "-no-CApath" option for
the same effect.

In addition because of the fact that validation errors are not treated
as command errors I forgot to check for command errors.